Add feature - Export Favorites list as text
Added ability to export favorites list as text. Available under the
Advanced settings. Exports in a format like this:
```
627 - Tech Support Cheat Sheet
641 - Free
645 - RPS
675 - Revolutionary
688 - Self-Description
```

The user can choose any app to export to that supports text/plain, such
as Gmail or a note-taking app.

Note that my computer and phone are not set up for Android development,
so I haven't tested or even compiled this, but it's a pretty simple
feature and hopefully I have done everything right. I used google
translate to add new text to the other languages, but I cannot vouch for
their accuracy.