Fix NullPointerException
At the end of the file, `it` would be null and trying to assign it to `line: String!` throws and exception:

> java.lang.NullPointerException: it must not be null

The exception is caught, and since it is at the end it does not impact the import.
Still, let's fix it.

Need to pull out the setFavourite because calling a coroutine within forEachLine does not work.